Correction of "1. Armazéns específicos"

+++ Generics_Exercises/armazem_especifico/Comida.java
@@ -1,9 +1,9 @@
 public class Comida {
     private String nome;
-    private int calorias;
+    private double calorias;
     private double preco;
 
-    public Comida(String nome, int calorias, double preco) {
+    public Comida(String nome, double calorias, double preco) {
         this.nome = nome;
         this.calorias = calorias;
         this.preco = preco;
@@ -11,6 +11,6 @@ public class Comida {
 
     @Override
     public String toString() {
-        return String.format("[%s] %d R$ %.1f", nome, calorias, preco);
+        return String.format("[%s] %.6f R$ %.6f", nome, calorias, preco);
     }
 }
